Add support for not enforcing XFS quotas.
Add XFS disk isolator support for not enforcing disk quotas on
containers. While there is a global filesystem configuration option
to turn off quota enforcement, we should not automatically toggle
that because we don't know why the operator might have changed that
configuration. Instead, we use the soft quota limit to implement
an accounting quota and the hard limit for an enforced quota.

Review: https://reviews.apache.org/r/55897