Optionally isolate only the agent network ports.
Normally, the `network/ports` isolator will kill any task that
listens on a port that it does not have resources for. However,
executors that are based on the libprocess API will always listen
on a port in the ephemeral range, and we want to make it possible
to use libprocess-based executors.

Added the `--container_ports_watch_resources_only` option to only
kill tasks when they listen on un-allocated ports within the port
range published by the agent resources. This still prevents port
collisions between tasks, but doesn't kill them just because the
executor is listening on a port.

Review: https://reviews.apache.org/r/60591