Added `network/ports` isolator recovery tests.
Added a test for the `network/ports` isolator recovery by starting
a task that is listening on a rogue port. We only configure the
isolator when we restart the agent to simulate the case where a
task only starts to misbehave after an agent recovery.

Review: https://reviews.apache.org/r/60593