Stopped keeping multiple limitations in MesosContainerizer.
MesosContainerizer was keeping a vector of ContainerLimitation
objects, but in practice it was not possible to have more than one at
a time, since receiving a limitation initiates a container destroy
and new limitations are dropped once the destroy begins. We replace
the vector<ContainerLimition> with a Option<ContainerLimitation>,
but otherwise preserve all the existing semantics.