Added a `ns::supported` convenience API.
Added a `ns::supported` convenience API which directly expresses the
intent to probe whether a specific Linux namespace is supported on
the running kernel and takes care of kernel versioning special cases.

This API subsumes the previous usages of `ns::namespaces`, so that API
is now used only retained as an internal helper.

Review: https://reviews.apache.org/r/64226